Revert " add param network, add file unitTest, add fun getLatestBlockNumber, getDataOnNetWork"
