+++ .env.example
@@ -5,6 +5,8 @@ INFURA_WS_URL=wss://mainnet.infura.io/ws
 
 ETH_BLOCK_TIME=30
 
+TOKEN_CONTRACT_ADDRESS=0xb4e95c65c23cab8c4fcaf15e4936c392e6ab21bf
+
 WALLET_FROM=
 WALLET_TO=
 AMOUNT=
