+++ src/watcher.js
@@ -1,18 +1,14 @@
 const Web3 = require('web3')
-const contract_config = require('./config')
+const contract_abi = require('./abi/config')
 
+var admin = require("firebase-admin");
+var serviceAccount = require("./keys/serviceAccountKey.json");
 
-require('./env')
-
-var actionSideChain = require('../src/sideChain/actionSideChain');
-
-var db = {};
-
-setDataBase = (dataBase) => {
-  db = dataBase;
-  actionSideChain.setDataBase(db);
-}
-
+admin.initializeApp({
+  credential: admin.credential.cert(serviceAccount),
+  databaseURL: process.env.FIREBASE_DB_URL
+});
+var db = admin.database();
 
 writeEventListData = (col, datas) => {
   if(datas.length ==0)
@@ -26,7 +22,6 @@ writeEventData = (col, txHash, data) => {
   db.ref(col+ '/' + txHash).set(data);
 }
 
-
 getLatestBlock = (col, callback) => {
   return db.ref(col+ '/').orderByChild("blockNumber").limitToLast(1).once('value', function(snapshot){
       if(!snapshot.exists()){
@@ -34,79 +29,52 @@ getLatestBlock = (col, callback) => {
         return callback(process.env.FROM_BLOCK)
       }
       for(key in snapshot.val()){
-        //console.log('LatestBlock of ' + col + ' is: '+ snapshot.val()[key].blockNumber)
+        console.log('LatestBlock of ' + col + ' is: '+ snapshot.val()[key].blockNumber)
         return callback(snapshot.val()[key].blockNumber)
       }
   })
 }
 
-watchEvent = (contractName, eventName, network) => {
+watchEvent = (contractName, eventName) => {
   // Instantiate web3 with WebSocketProvider
   const web3 = new Web3(new Web3.providers.WebsocketProvider(process.env.INFURA_WS_URL))
-  if(!contract_config[network][contractName])
+  if(!contract_abi[contractName])
     return
   // Instantiate token contract object with JSON ABI and address
 
   const tokenContract = new web3.eth.Contract(
-      contract_config.json[contractName].abi, contract_config[network][contractName].address,
+    contract_abi[contractName].abi,contract_abi[contractName].address,
     (error, result) => { if (error) console.log(error) }
   )
   let options = {
   }
   // Subscribe to Transfer events matching filter criteria
-  tokenContract.events[eventName](options,function(error, event){ 
-      if(error) {
-        console.log('error', error);
-      } else {
-      }
+  tokenContract.events[eventName](options
+    , function(error, event){ 
+      console.log(event); 
   })
   .on('data', function(event){
-      
-      console.log(contractName + ' ' + eventName + ' at ' + new Date() + ' line 70 txHash ->  ' + event.transactionHash); 
-
-      let data = event.returnValues;
-      data.blockNumber = event.blockNumber;
-      data.txHash = event.transactionHash;
-      data.mintStage = 0;
-      writeEventData(contractName + '/' + eventName, event.transactionHash, data);
-
-      let objSwap = JSON.parse(process.env.TOKEN_SWAP);
-
-      if(contractName == 'BBWrap' && eventName == 'DepositEther') {
-          let keyToken = objSwap.ETHER;
-          if(keyToken === undefined || keyToken == '') {
-            console.log('error : process.env.TOKEN_SWAP.ETHER');
-          } else {
-            actionSideChain.mintToken('BBWrap/DepositEther', keyToken);
-          }
-      } else if(contractName == 'BBWrap' && eventName == 'DepositToken') {
-          let token = data.token.toLowerCase();
-          
-          let keyToken = objSwap.TOKEN[token];
-          if(keyToken === undefined || keyToken == '') {
-            console.log('error : process.env.TOKEN_SWAP.TOKEN. ',token);
-          } else {
-            actionSideChain.mintToken('BBWrap/DepositToken', keyToken);
-          }
-      } else if(contractName == 'BBWrap' && eventName == 'MintToken') {
-          actionSideChain.mintToken('BBWrap/DepositEther');
-      }
+      let data = event.returnValues
+      data.blockNumber = event.blockNumber
+      data.txHash = event.transactionHash
+      writeEventData(contractName + '/' + eventName, event.transactionHash, data)
   })
   .on('changed', function(event){
       // remove event from local database
   })
   .on('error', console.error);
 }
-
-getPastEvents = (contractName, eventName, network) => {
+getPastEvents = (contractName, eventName) => {
   // Instantiate web3 with WebSocketProvider
   const web3 = new Web3(new Web3.providers.HttpProvider(process.env.INFURA_URL))
 
-  if(!contract_config[network][contractName])
+  // Instantiate token contract object with JSON ABI and address
+
+  if(!contract_abi[contractName])
     return
   //todo
   const tokenContract = new web3.eth.Contract(
-    contract_config.json[contractName].abi,contract_config[network][contractName].address,
+    contract_abi[contractName].abi,contract_abi[contractName].address,
     (error, result) => { if (error) console.log(error) }
   )
   return getLatestBlock(contractName + '/' + eventName, function(blockNumber){
@@ -115,10 +83,7 @@ getPastEvents = (contractName, eventName, network) => {
     let options = {
       fromBlock: parseFloat(blockNumber)+1
     }
-    return tokenContract.getPastEvents(eventName, options , 
-      function(error, events){ 
-        console.log(events); 
-      })
+    return tokenContract.getPastEvents(eventName, options , function(error, events){ console.log(events); })
     .then(function(events){
         let datas = []
         for (var i = events.length - 1; i >= 0; i--) {
@@ -128,15 +93,12 @@ getPastEvents = (contractName, eventName, network) => {
           data.txHash = events[i].transactionHash
           datas.push(data)
         }
-        console.log('writeEventListData');
-        return writeEventListData('BBOToken' + '/' + eventName, datas);
+        return writeEventListData(contractName + '/' + eventName, datas);
     });
   })
   
 }
 module.exports = {
-  watchEvent,
-  writeEventData, 
-  getPastEvents,
-  setDataBase
+  watchEvent, 
+  getPastEvents
 }
