--- test/contractAddress.json
@@ -1,1 +1,0 @@
-{"BBOTest":{"address":"0x474B1870A7506Be3ecc144ccaA215c1296521D88","events":["Transfer"]},"JOB":{"address":"0x8B0Fd2Ab71C879de67AE66B443FC5840310e1B4F","events":["JobCreated","JobCanceled","JobStarted","JobFinished"]},"BID":{"address":"0xBAD87A43803b035264d159C9F15Af4ac321046f8","events":["BidCreated","BidCanceled","BidAccepted"]},"RATING":{"address":"0x29CfDD6d056Bf1CA7c21b9A9751326A50dd8cac7","events":["Rating"]},"PAYMENT":{"address":"0xbB2C27B6f434f0228041a947c20b92c34EA87818","events":["PaymentClaimed","PaymentAccepted","PaymentRejected","DisputeFinalized","DepositMoney"]},"DISPUTE":{"address":"0x523A2646A5Ea7e2009Aa6576ceA13fC90934c368","events":["DisputeFinalized","DisputeStarted","DisputeAgainsted","DisputeUpdated"]},"VOTING":{"address":"0x47530468EF5Aaccfe8007cC2E6DAAB64eafbe44C","events":["PollStarted","PollUpdated","PollOptionAdded","VotingRightsGranted","VotingRightsWithdrawn","VoteCommitted","VoteRevealed"]},"VOTINGHELPER":{"address":"0x52fbb491eC538C69B49D205d2F4689E559C489c9","events":[]}}
