--- testHelper/utils/encodeCall.js
@@ -1,17 +1,0 @@
-var abi = require('ethereumjs-abi')
-var BN = require( 'bignumber.js')
-
-function formatValue(value) {
-  if (typeof(value) === 'number' || BN.isBigNumber(value)) {
-    return value.toString();
-  } else {
-    return value;
-  }
-}
-
-function encodeCall(name, args = [], rawValues = []) {
-  const values = rawValues.map(formatValue)
-  const methodId = abi.methodID(name, args).toString('hex');
-  const params = abi.rawEncode(args, values).toString('hex');
-  return '0x' + methodId + params;
-}
